server/requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.50.0
idna==3.13
psycopg==3.3.4
psycopg-binary==3.3.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
